+++ lab-3/src/main/java/ru/akirakozov/sd/refactoring/servlet/AddProductServlet.java
@@ -20,7 +20,7 @@ public class AddProductServlet extends HttpServlet {
     @Override
     protected void doGet(HttpServletRequest request, HttpServletResponse response) throws IOException {
         String name = request.getParameter("name");
-        long price = Long.parseLong(request.getParameter("price"));
+        int price = Integer.parseInt(request.getParameter("price"));
 
         productDB.addProduct(new Product(name, price));
